# QuickStart/Fungi_api_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fungi API",
    "description": "API for Fungi project",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development"
    }
  ],
  "paths": {
...
    "/Users/GetUser": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user",
        "description": "Receives information about the user by id",
        "operationId": "GetUser",
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Users/GetFilteredUsers": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get filtered users",
        "description": "Gets a list of users with filters",
        "operationId": "GetFilteredUsers",
        "parameters": [
          {
            "name": "PartOfUsername",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PartOfEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "RoleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetFilteredUsersResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFilteredUsersResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFilteredUsersResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Users/CreateUser": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "description": "Creates user",
        "operationId": "CreateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/Users/UploadMyAvatar": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Upload current user avatar",
        "description": "Uploads avatar image, crops it to square and stores as WebP",
        "operationId": "UploadMyAvatar",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Avatar"
                ],
                "type": "object",
                "properties": {
                  "Avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Avatar": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UploadMyAvatarResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadMyAvatarResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadMyAvatarResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/Users/UpdateUser": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "description": "Updates user info by id",
        "operationId": "UpdateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Users/DeleteUser": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "description": "Deletes user by id",
        "operationId": "DeleteUser",
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponseBaseResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponseBaseResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponseBaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Users/UpdateUserSmallParam": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Update user with small parameters",
        "description": "Short version of UpdateUser",
        "operationId": "UpdateUserSmallParam",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "NewName": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "NewName": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string"
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "CreateUserResponse": {
        "type": "object",
        "properties": {
          "createdUserId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "CreateUserResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateUserResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "DeleteUserResponse": {
        "type": "object",
        "properties": {
          "deletedUserId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "DeleteUserResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeleteUserResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "ExceptionDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string"
          },
          "errorGroup": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "GetFilteredUsersResponse": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          }
        },
        "additionalProperties": false
      },
      "GetFilteredUsersResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GetFilteredUsersResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "additionalProperties": false
      },
      "GetUserResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GetUserResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "accessLevel": {
            "type": "integer",
            "format": "int32"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "newUsername": {
            "type": "string"
          },
          "newEmail": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string"
          },
          "newRoleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "UpdateUserResponse": {
        "type": "object",
        "properties": {
          "updatedUserId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "UpdateUserResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UpdateUserResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
...
      "UploadMyAvatarResponse": {
        "type": "object",
        "properties": {
          "avatarUrl": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UploadMyAvatarResponseBaseResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UploadMyAvatarResponse"
          },
          "errorMessage": {
            "$ref": "#/components/schemas/ExceptionDto"
          }
        },
        "additionalProperties": false
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/RoleDto"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}